Minor bugfix, minor enhancements to debug text, and change to allow arbitrary phdr relocation (#2)
* Minor bugfix

Method 2 for inserting a new segment removed the GNU_STACK phdr
entry without decreasing e_phnum, which is later dumped in the
elf header.  If method 2 is used and it does not decrease e_phnum,
then the loader will think there is one more segment than there
really is.

* Changed error to warning to allow for overwriting .text with phdrs

This change allows a binary with no space for additional phdrs to
move the phdrs to overwrite the start of the .text section.  This
obviously requires runtime support from the binary to restore the
original .text contents, so this library does not provide a
function to overwrite .text contents.  Rather, it simply allows
code using it as a library to relocate the phdrs anywhere, even if
the new location overlaps another segment.